// src/interfaces/visitation-information.interface.ts
export interface VisitationInformation {
  visitation_id: string;
  tg_id: string;
  date_logged: string;
  comment: string;
  presence_flag: number;
  member_verified_flag: number;
  app_version: string;
  imei: string;
  staff_id: string;
  phone_number: string;
  latitude: string;
  longitude: string;
  image_name: string;
}